feat #936: Hallazgos
- Refactor consulta de entradas.
